Migrate all parameters to a common infrastructure

+++ subprojects/gradle-fixtures/src/main/java/dev/gradleplugins/test/fixtures/gradle/executer/internal/parameters/WelcomeMessageParameter.java
@@ -1,8 +1,6 @@
 package dev.gradleplugins.test.fixtures.gradle.executer.internal.parameters;
 
-import lombok.Value;
 import lombok.val;
-import org.apache.commons.io.FileUtils;
 import org.gradle.launcher.cli.DefaultCommandLineActionFactory;
 import org.gradle.util.GradleVersion;
 
@@ -11,46 +9,34 @@ import java.io.UncheckedIOException;
 import java.util.Collections;
 import java.util.Map;
 
-public interface WelcomeMessageParameter extends JvmSystemPropertyParameter {
-    void apply(GradleUserHomeDirectoryParameter gradleUserHomeDirectory, GradleVersion version);
-
-    static WelcomeMessageParameter enabled() {
-        return new ShowingWelcomeMessageParameter();
-    }
-
-    static WelcomeMessageParameter disabled() {
-        return new HidingWelcomeMessageParameter();
-    }
-
-    @Value
-    class ShowingWelcomeMessageParameter implements WelcomeMessageParameter {
-        @Override
-        public Map<String, String> getAsJvmSystemProperties() {
-            return Collections.singletonMap(DefaultCommandLineActionFactory.WELCOME_MESSAGE_ENABLED_SYSTEM_PROPERTY, Boolean.TRUE.toString());
-        }
-
-        @Override
-        public void apply(GradleUserHomeDirectoryParameter gradleUserHomeDirectory, GradleVersion version) {
+public final class WelcomeMessageParameter extends GradleExecutionParameterImpl<WelcomeMessage> implements JvmSystemPropertyParameter<WelcomeMessage>, GradleExecutionParameter<WelcomeMessage> {
+    public void apply(GradleUserHomeDirectory gradleUserHomeDirectory, GradleVersion version) {
+        if (get().equals(WelcomeMessage.ENABLED)) {
             val welcomeMessageFile = gradleUserHomeDirectory.file("notifications/" + version.getVersion() + "/release-features.rendered");
             welcomeMessageFile.delete();
-        }
-    }
-
-    @Value
-    class HidingWelcomeMessageParameter implements WelcomeMessageParameter {
-        @Override
-        public Map<String, String> getAsJvmSystemProperties() {
-            return Collections.singletonMap(DefaultCommandLineActionFactory.WELCOME_MESSAGE_ENABLED_SYSTEM_PROPERTY, Boolean.FALSE.toString());
-        }
-
-        @Override
-        public void apply(GradleUserHomeDirectoryParameter gradleUserHomeDirectory, GradleVersion version) {
+        } else {
             val welcomeMessageFile = gradleUserHomeDirectory.file("notifications/" + version.getVersion() + "/release-features.rendered");
             try {
-                FileUtils.touch(welcomeMessageFile);
+                welcomeMessageFile.touch();
             } catch (IOException e) {
                 throw new UncheckedIOException("Could not ensure render message is properly rendered", e);
             }
         }
+    }
+
+    public static WelcomeMessageParameter enabled() {
+        return fixed(WelcomeMessageParameter.class, WelcomeMessage.ENABLED);
+    }
+
+    public static WelcomeMessageParameter disabled() {
+        return fixed(WelcomeMessageParameter.class, WelcomeMessage.DISABLED);
+    }
+
+    @Override
+    public Map<String, String> getAsJvmSystemProperties() {
+        if (get().equals(WelcomeMessage.ENABLED)) {
+            return Collections.singletonMap(DefaultCommandLineActionFactory.WELCOME_MESSAGE_ENABLED_SYSTEM_PROPERTY, Boolean.TRUE.toString());
+        }
+        return Collections.singletonMap(DefaultCommandLineActionFactory.WELCOME_MESSAGE_ENABLED_SYSTEM_PROPERTY, Boolean.FALSE.toString());
     }
 }
